// Movie-List/src/Types/IMovie.ts
export default interface IMovie {
  Poster: string;
  Title: string;
  Type: string;
  Year: string;
  imdbID: string;
  // [propName: string]: any;
  Released?: string;
  Genre?: string;
  Director?: string;
  Actors?: string;
  Plot?: string;
  Language?: string;
  Awards?: string;
  imdbRating?: string;
  DVD?: string;
  BoxOffice?: string;
  totalSeasons?: string;
  Response?: string;
}